Move CardPresentPaymentsPlugins to its own file
